Fix to not mark a valid preloaded subtasks-file as INVALID
Before this change, a subtasks-file was marked as INVALID when
a task had a preload-task pointing to a task from external file which
had a subtasks-file [1]. While preloading a task, all the subtasks
are copied to the current task. Since the subtasks-file property was
a String, it does not maintain the information regarding the location
of the subtasks-file. Due to which, the plugin fails to lookup
the subtasks-file.

Thus, fix this issue by using FileKey as the type for subtasks-file
property, which helps in tracking the location of the subtasks-file.
Add tests for the same.

[1]
file: `All-Projects:refs/meta/config:task.config`
```
[root "root task"]
    applicable = status:open
    preload-task = %sample group^Sample task with subtasks-file
```

file: `All-Users:refs/groups/<uuid of sample group>:task.config`
```
[task "Sample task with subtasks-file"]
  applicable = is:open
  pass = is:open
  subtasks-file = bar.config
```

file: `All-Users:refs/groups/<uuid of sample group>:task/bar.config`
```
[task "Sample task"]
  applicable = is:open
  pass = True
```

Change-Id: Icb1c548655eace95fefa2093bdc827276e0908d8