UI: Move tasks to a primary tab
Tasks are now rendered in a primary tab. This makes the UI compact and
consistent. The API to fetch tasks has been moved from GrTaskPlugin to
GrTaskSummary because the summary is registered to the commit-container
which renders by default when the change page is loaded. The tasks tab
content is attached to 'change-view-tab-content' which only renders when
the tab is clicked or focus is changed to it.

Screenshots: https://imgur.com/a/Qr2i1pQ

Change-Id: I14a21fc83ef04a909576dfd796624044cbf72858