Update to 3.9.1
Update build/test files and bazlets to bring in a newer rules_nodejs
version. Also update the bazel version and eslintrc to match core
stable-3.9.

The tests require adding '--add-opens java.base/java.lang=ALL-UNNAMED'
to the container.javaOptions because of the reflection used in
com.googlesource.gerrit.plugins.task.properties.

The `lint_test` build rule for ESLint doesn't seem to have a way to work
for both in-tree and standalone execution, so update it to continue
passing in-tree since that's how CI runs it.

Change-Id: Ibded565ba8ac98b301e1b4285b6182804df5b4f4