test: docker: Update to Alpine 3.16
Alpine 3.11 has been EOL since 2021-11-01 and 3.16 is the oldest version
that's still supported. This upgrade requires two fixes:
  1) Moving to Python 3
  2) Working around the OpenSSH 9 scp backwards-incompatible API change

For #1, the parent changes removed the inline Python usage and then
updated the two Python scripts used by tests to support both Python 2
and 3. Modifying their shebang lines to explicitly use `python3` is all
that's needed in this change.

For #2, the solution adopted by upstream was to always prefer using curl
to download the commit-msg hook script instead of scp. Using that for
the tests is a simple fix.

Change-Id: I4ea8dfdf98ace06918ee6d680b8a3b77f7523601