test: Consistently apply json_pp to expected
We use json_pp on both expected and actual results to ensure the
comparisons are equal. Some expected files weren't getting json_pp
applied because it was missed or the file was reused for more processing
and then didn't get updated. Fix all expected files to use json_pp.

This also allows us to later change the internals of json_pp and keep
the tests passing.

Change-Id: I24673890beb9e687e39d11115e7b063ff2c286e3