Add support to skip evaluation of expensive tasks
This change introduces a new task config key i.e. `evaluation-threshold`
to allow skipping evaluation of expensive tasks when the number of
changes in the gerrit query output (or createPluginDefinedInfos() input)
is more than the value defined for this key.

Since task plugin UI only supports passing a single change to
`createPluginDefinedInfos()`, it'll never produce `SKIPPED` tasks and
thus doesn't need to be updated.

Change-Id: I333058472ca75b27a6bc64549b5a35a45ca0389f