Create an 'extensions' subpackage
This will allow other plugins providing task name factory
implementations a way to include the build dependency through Bazel.

Since standalone and in-tree builds will have a different absolute
package path for the new subpackage, use a bit of code from the
bazel-skylib repo to do the equivalent of os.path.join() with the
package name.

Change-Id: Id35793fc73dbf96a1e597c2552769c8469175634