Fix duplicate detection for tasks generated by change tasks-factory
When both the `--task--applicable` and `--task--invalid` attribute
flags are passed to a gerrit query, the task plugin currently fails to
detect duplicate tasks generated by change type tasks-factory. This
happens because the `nodeSetByBaseTasksFactory` map is not populated.

As a result, if the change tasks-factory produces duplicate tasks, such
as in the case when the query used for generating the tasks results in
cyclic dependencies, the plugin attempts to create nodes for each task.
This leads to excessive memory usage, degraded query performance, and
in extreme cases, server instability.

This change fixes the issue by ensuring the `nodeSetByBaseTasksFactory`
map is correctly populated.

Change-Id: If747f9ff1f4ee807f9630de2974b159cee4a762f