Fix duplicate detection for tasks generated by tasks-factory
When both the `--task--applicable` and `--task--invalid` attribute
flags are passed to a gerrit query, the task plugin currently fails to
detect duplicate tasks generated by tasks-factory. This happens because
the `nodeSetByBaseTasksFactory` map is not populated.

As a result, when a change-type task-factory is defined in the task
config, it may generate a large number of duplicate change tasks,
causing high memory usage on the server and slower response times for
queries.

Fix the issue by ensuring the `nodeSetByBaseTasksFactory` map is
correctly populated.

Change-Id: If747f9ff1f4ee807f9630de2974b159cee4a762f